requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
ecdsa==0.19.0
fastapi==0.115.5
firebase-admin==6.6.0
gcloud==0.18.3
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
h11==0.14.0
httplib2==0.22.0
idna==3.10
jwcrypto==1.5.6
msgpack==1.1.0
oauth2client==4.1.3
proto-plus==1.25.0
protobuf==5.29.0rc3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.0
pyparsing==3.2.0
Pyrebase4==4.8.0
python-jose==3.3.0
python-jwt==4.1.0
python-multipart==0.0.17
requests==2.32.3
requests-toolbelt==0.10.1
rsa==4.9
setuptools==75.5.0
six==1.16.0
sniffio==1.3.1
starlette==0.41.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.32.0
wheel==0.45.0